pyproject.toml
[project]
name = "azure-ml-finetuning-eval-skills"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "azure-ai-evaluation>=1.13.7",
    "azure-ai-ml>=1.30.0",
    "azure-identity>=1.25.1",
    "openai>=2.9.0",
    "python-dotenv>=1.2.1",
    "wikipedia>=1.4.0",
]
